Adds camera to scene and ability to fly around

+++ seagull/headers/mesh.h
@@ -67,7 +67,7 @@ class Mesh {
                 else if(name == "texture_normal")
                     number = std::to_string(normalNr++);
 
-                shader.setFloat(("material." + name + number).c_str(), i);
+                shader.setInt(("material." + name + number).c_str(), i);
                 GLCall(glBindTexture(GL_TEXTURE_2D, textures[i].id));
             }
